--- Samples/Swift/nhubsample/nhubsample/Constants.swift
@@ -1,12 +1,0 @@
-//
-//  Copyright © 2018 Microsoft All rights reserved.
-//  Licensed under the Apache License (2.0).
-//
-
-import Foundation
-
-struct Constants {
-    static let NHInfoConnectionString = "NotificationHubConnectionString";
-    static let NHInfoHubName = "NotificationHubName";
-    static let NHUserDefaultTags = "notification_tags";
-}
