+++ Samples/Swift/nhubsample/nhubsample/ViewController.swift
@@ -4,6 +4,7 @@
 //
 
 import UIKit
+import UserNotifications
 
 class ViewController: UIViewController {
     @IBOutlet weak var tagsTextField: UITextField!
@@ -12,30 +13,56 @@ class ViewController: UIViewController {
     
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         // Simple method to dismiss keyboard when user taps outside of the UITextField.
-        self.view.endEditing(true)
+        view.endEditing(true)
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-        
-        self.tagsTextField.text = UserDefaults.standard.string(forKey: Constants.NHUserDefaultTags)
+        setupUI()
     }
+    
+}
 
+private extension ViewController {
+    
+    func setupUI() {
+        tagsTextField.text = UserDefaults.standard.string(forKey: Constants.NHUserDefaultTags)
+    }
+    
+    func getNotificationSettings() {
+        UNUserNotificationCenter.current().getNotificationSettings { settings in
+            print("Notification settings: \(settings)")
+            guard settings.authorizationStatus == .authorized else { return }
+            DispatchQueue.main.async {
+                UIApplication.shared.registerForRemoteNotifications()
+            }
+        }
+    }
+    
     @IBAction func handleRegister(_ sender: Any) {
         // Save raw tags text in storage
-        UserDefaults.standard.set(self.tagsTextField.text, forKey: Constants.NHUserDefaultTags)
+        UserDefaults.standard.set(tagsTextField.text, forKey: Constants.NHUserDefaultTags)
         
-        //
-        // Delegate processing the register action to the app delegate.
-        //
-        let appDelegate: AppDelegate? = UIApplication.shared.delegate as? AppDelegate
-        appDelegate?.handleRegister()
+        UNUserNotificationCenter.current()
+            .requestAuthorization(options: [.alert, .sound, .badge]) {
+                granted, error in
+                if (error != nil) {
+                    print("Error requesting for authorization:");
+                }
+                print("Permission granted: \(granted)")
+                guard granted else { return }
+                self.getNotificationSettings()
+        }
     }
     
     @IBAction func handleUnregister(_ sender: Any) {
-        let appDelegate: AppDelegate? = UIApplication.shared.delegate as? AppDelegate
-        appDelegate?.handleUnregister()
+        let hub = getNotificationHub()
+        hub.unregisterNative { error in
+            if (error != nil) {
+                print("Error unregistering for push: \(error.debugDescription)");
+            } else {
+                showAlert("Unregistered", withTitle: "Registration Status")
+            }
+        }
     }
-    
 }
